requirements.txt
django==1.7.2
django-crispy-forms==1.4.0
oauthlib==0.7.2
requests==2.5.1
wsgiref==0.1.2
yboss==0.1.1
